+++ example/lib/barcode_scanner_pageview.dart
@@ -69,7 +69,7 @@ class _BarcodeScannerPage extends StatelessWidget {
         MobileScanner(
           controller: controller,
           fit: BoxFit.contain,
-          errorBuilder: (context, error, child) {
+          errorBuilder: (context, error) {
             return ScannerErrorWidget(error: error);
           },
         ),