+++ example/lib/barcode_scanner_returning_image.dart
@@ -87,7 +87,7 @@ class _BarcodeScannerReturningImageState
                   children: [
                     MobileScanner(
                       controller: controller,
-                      errorBuilder: (context, error, child) {
+                      errorBuilder: (context, error) {
                         return ScannerErrorWidget(error: error);
                       },
                       fit: BoxFit.contain,