+++ example/lib/barcode_scanner_window.dart
@@ -39,7 +39,7 @@ class _BarcodeScannerWithScanWindowState
             fit: boxFit,
             scanWindow: scanWindow,
             controller: controller,
-            errorBuilder: (context, error, child) {
+            errorBuilder: (context, error) {
               return ScannerErrorWidget(error: error);
             },
           ),