+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * The `updateScanWindow` method is now private. Instead, update the scan window in the `MobileScanner` widget directly.
 * The deprecated `EncryptionType.none` constant has been removed. Use `EncryptionType.unknown` instead.
+* The `errorBuilder` and `placeholderBuilder` of the `MobileScanner` widget no longer take a Widget argument, as it was unused.
+* The `MobileScannerErrorBuilder` typedef has been removed.
 
 Bugs fixed:
 * [Apple] Fixed an issue which caused the scanWindow to always be present, even when reset to no value.